% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files_and_metadata-utils.R
\name{get_all_files_recursive}
\alias{get_all_files_recursive}
\title{Get all project's files}
\usage{
get_all_files_recursive(path)
}
\arguments{
\item{path}{Project files directory path.}
}
\description{
Utility function for fetching all project files recursively from SB File
 system within Data Studio.
}
