% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_save_file_generic.R
\name{mod_save_file_generic_ui}
\alias{mod_save_file_generic_ui}
\title{save_file_generic UI Function}
\usage{
mod_save_file_generic_ui(id, save_button_title = "Save file")
}
\arguments{
\item{id}{Module ID.}

\item{save_button_title}{Button label.}
}
\value{
No value. Use in the UI & server of shiny application.
}
\description{
A shiny Module.
}
