% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_save_plot_for_export.R
\name{save_plot_for_export_server}
\alias{save_plot_for_export_server}
\title{Save `plot` for export module server function}
\usage{
save_plot_for_export_server(
  id,
  plot_rv,
  output_formats = c("png", "pdf", "svg", "jpeg", "bmp", "tiff"),
  module_title = NULL,
  sbg_directory_path = "/sbgenomics"
)
}
\arguments{
\item{id}{Module's ID.}

\item{plot_rv}{A `reactiveValues` variable with a slot `plot` containing an
object created with `recordPlot()` function you would like to save.}

\item{output_formats}{Output formats offered to the user. Can be a subset of:
"png", "pdf", "svg", "jpeg", "bmp", or "tiff".}

\item{module_title}{Title (top left corner) of a modal (popup window) with
settings.}

\item{sbg_directory_path}{Path to the sbgenomics directory containing
project-files, output-files and workspace subdirectories on the instance.
These directories are expected to exist on the instance where the app would
run. For the purposes of testing your app locally, you can create a mock
directory `sbgenomics` with the same structure - containing sub-directories
`project-files`, `output-files` and `workspace`.}
}
\value{
No value. Use in UI & server of shiny application.
}
\description{
A server side of a shiny Module that allows users to save plots
 generated within the Shiny app and export them to the project on the
 Seven Bridges Platform.
}
