% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_save_plot_for_export.R
\name{save_plot_for_export_ui}
\alias{save_plot_for_export_ui}
\title{Save `plot` for export module UI function}
\usage{
save_plot_for_export_ui(id, save_button_title = "Save plot")
}
\arguments{
\item{id}{Module ID.}

\item{save_button_title}{Button label.}
}
\value{
No value. Use in UI & server of shiny application.
}
\description{
UI part of the module that can be used to save the `plot`
 for export to Seven Bridges Platform. The UI contains a button that triggers
 a modal dialog with further settings.
}
