% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_save_plot_for_export.R
\name{save_plot_modalDialog_ui}
\alias{save_plot_modalDialog_ui}
\title{`save_plot_for_export` module UI function}
\usage{
save_plot_modalDialog_ui(
  id,
  output_formats,
  ns,
  initial_plot_width = 868,
  initial_plot_height = 400
)
}
\arguments{
\item{id}{Module ID.}

\item{output_formats}{Output formats offered to the user. Can be a subset of:
"png", "pdf", "svg", "jpeg", "bmp", or "tiff".}

\item{ns}{Namespace function for handling input IDs.}

\item{initial_plot_width}{Integer representing the initial width of the plot
(the default value is 868).}

\item{initial_plot_height}{Integer representing the initial height of the
plot (the default value is 400).}
}
\value{
No value. Use in UI & server of shiny application
}
\description{
UI function of the module for saving plots for export to the
 Seven Bridges Platform. The function adds an action button that triggers
 a plot export modal. This module allows users to adjust the plot they wish
 to save to Seven Bridges Platform, by resizing it and then choosing the
 file format for export: png, pdf, svg, jpeg, bmp, or tiff.
 The UI function should be placed in the UI part of the Shiny app
 where the file export button is required.
 Please, have in mind that this module only saves the plot, while the actual
 export of the saved plots happens when the users stop the app.
}
