% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files_and_metadata-utils.R
\name{get_all_project_files}
\alias{get_all_project_files}
\title{Get all project's files}
\usage{
get_all_project_files(path)
}
\arguments{
\item{path}{Project files directory path.}
}
\value{
Data.frame containing all project files with their names, paths,
 sizes and associated metadata fields if available.
}
\description{
Utility function for fetching all project files recursively from SB File
 system within Data Studio along side with their metadata associated if
 available. The expected output is the data.frame containing all files
 information available, ready to be used and displayed within file pickers.
}
