% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_picker.R
\name{mod_file_picker_server}
\alias{mod_file_picker_server}
\title{file_picker Server Functions}
\usage{
mod_file_picker_server(
  id,
  files_df,
  selection = "single",
  file_identifier_column = "path",
  ...
)
}
\arguments{
\item{id}{A unique identifier for the module instance.}

\item{files_df}{A data frame containing file information. This data frame
should have a column for file paths and any other relevant metadata. You
can use the \code{\link{get_all_project_files}} function to fetch all
project files along with their metadata from the SB File system
within Data Studio.This function returns a data frame containing
comprehensive file information, making it an ideal input for the
`mod_file_picker_server()` function.}

\item{selection}{A string specifying the selection mode. Can be either
'single' for single file selection or 'multiple' for multiple file
 selection. Defaults to 'single'.}

\item{file_identifier_column}{A string specifying the column name in
`files_df` from which the values of selected files will be returned.
 Defaults to `'path'`.}

\item{...}{Additional parameters to be passed to the `reactable()` function
this module relies on}
}
\value{
A reactive expression containing information about the selected
 files based on the specified `file_identifier_column`.
}
\description{
Server function of a Shiny module for selecting either a single
 file or multiple files to be used within the app. This module displays a
 modal dialog with a preview table of files and allows users to select files
 based on the specified `selection` mode ('single' or 'multiple').
}
\details{
To incorporate this module into your Shiny app, you need to include
 both the UI and server functions in the appropriate places in your app
 code:
 \itemize{
   \item **UI Function**: Add `mod_file_picker_ui("file_picker_1")` to the
    UI part of your Shiny app where you want the action button to appear.
   \item **Server Function**: Add `mod_file_picker_server("file_picker_1", files_df)`
    to the server part of your Shiny app. Make sure to replace `files_df`
    with your actual data frame containing file information.
 }
}
\examples{
library(shiny)
library(sbShinyModules)

# App's  UI
ui <- fluidPage(
  titlePanel("File Picker Module Examples"),
  sidebarLayout(
    sidebarPanel(
      # Single File Picker - UI
      fluidRow(
        h3("Single File Picker"),
        br(),
        sbShinyModules::mod_file_picker_ui("single_file_picker"),
        br(),
        h5("Selected File"),
        verbatimTextOutput("single_file_picker_selection")
      ),
      hr(),
      # Multiple Files Picker - UI
      fluidRow(
        h3("Multiple Files Picker"),
        br(),
        sbShinyModules::mod_file_picker_ui("multiple_files_picker"),
        br(),
        h5("Selected Files"),
        verbatimTextOutput("mult_files_picker_selection")
      )
    ),
    mainPanel(
      # Placeholder
    )
  )
)

# App's Server Logic
server <- function(input, output, session) {
  # ----------------------------- Load Files ----------------------------------
  # Load a built-in data frame for files
  files_df <- sbShinyModules::file_picker_example_data

  # Or fetch files from a provided directory (path) using the
  # get_all_project_files() utility function

  # files_df <- sbShinyModules::get_all_project_files(
  #   path = "/sbgenomics/project-files"
  # )

  # ---------------------------------------------------------------------------

  ## -------------------- Single File Picker - Server Code ---------------------

  # Call the file picker module
  selected_files_single_picker <- sbShinyModules::mod_file_picker_server("single_file_picker", files_df, selection = "single") # nolint

  # Display selected files
  output$single_file_picker_selection <- renderPrint({
    validate(
      need(
        selected_files_single_picker(),
        "No file has been selected."
      )
    )
    cat(selected_files_single_picker(), sep = "\n")
  })
  # ---------------------------------------------------------------------------



  ## ----------------- Multiple Files Picker - Server Code --------------------

  # Call the file picker module
  selected_files_mult_picker <- sbShinyModules::mod_file_picker_server("multiple_files_picker", files_df, selection = "multiple") # nolint

  # Display selected files
  output$mult_files_picker_selection <- renderPrint({
    validate(
      need(
        selected_files_mult_picker(),
        "No files have been selected."
      )
    )
    cat(selected_files_mult_picker(), sep = "\n")
  })
  # ---------------------------------------------------------------------------
}

# Run the Shiny app
shinyApp(ui, server)
}
\seealso{
\code{\link{mod_file_picker_ui}} for the corresponding server
 part of the module.
}
