% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_picker.R
\name{mod_file_picker_ui}
\alias{mod_file_picker_ui}
\title{file_picker UI Function}
\usage{
mod_file_picker_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
UI function of the file picker Shiny module. The function adds
 an action button that triggers a file picker modal. This module allows
 users to select either a single file or multiple files to be used within
 the app. The UI function should be placed in the UI part of the Shiny app
 where the file picker button is required.
}
\details{
To incorporate this module into your Shiny app, you need to include
 both the UI and server functions in the appropriate places in your app
 code:
 \itemize{
   \item **UI Function**: Add `mod_file_picker_ui("file_picker_1")` to the
    UI part of your Shiny app where you want the action button to appear.
   \item **Server Function**: Add `mod_file_picker_server("file_picker_1", files_df)`
    to the server part of your Shiny app. Make sure to replace `files_df`
    with your actual data frame containing file information.
 }
}
\examples{
library(shiny)
library(sbShinyModules)

# App's  UI
ui <- fluidPage(
  titlePanel("File Picker Module Examples"),
  sidebarLayout(
    sidebarPanel(
      # Single File Picker - UI
      fluidRow(
        h3("Single File Picker"),
        br(),
        sbShinyModules::mod_file_picker_ui("single_file_picker"),
        br(),
        h5("Selected Files"),
        verbatimTextOutput("single_file_picker_selection")
      ),
      hr(),
      # Multiple Files Picker - UI
      fluidRow(
        h3("Multiple Files Picker"),
        br(),
        sbShinyModules::mod_file_picker_ui("multiple_files_picker"),
        br(),
        h5("Selected Files"),
        verbatimTextOutput("mult_files_picker_selection")
      )
    ),
    mainPanel(
      # Placeholder
    )
  )
)

# App's Server Logic
server <- function(input, output, session) {
  # ----------------------------- Load Files ----------------------------------
  # Load a built-in data frame for files
  files_df <- sbShinyModules::file_picker_example_data

  # Or fetch files from a provided directory (path) using the
  # get_all_project_files() utility function

  # files_df <- sbShinyModules::get_all_project_files(
  #   path = "/sbgenomics/project-files"
  # )

  # ---------------------------------------------------------------------------

  ## -------------------- Single File Picker - Server Code ---------------------

  # Call the file picker module
  selected_files_single_picker <- sbShinyModules::mod_file_picker_server("single_file_picker", files_df, selection = "single") # nolint

  # Display selected files
  output$single_file_picker_selection <- renderPrint({
    validate(
      need(
        selected_files_single_picker(),
        "No file has been selected."
      )
    )
    cat(selected_files_single_picker(), sep = "\n")
  })
  # ---------------------------------------------------------------------------



  ## ----------------- Multiple Files Picker - Server Code --------------------

  # Call the file picker module
  selected_files_mult_picker <- sbShinyModules::mod_file_picker_server("multiple_files_picker", files_df, selection = "multiple") # nolint

  # Display selected files
  output$mult_files_picker_selection <- renderPrint({
    validate(
      need(
        selected_files_mult_picker(),
        "No files have been selected."
      )
    )
    cat(selected_files_mult_picker(), sep = "\n")
  })
  # ---------------------------------------------------------------------------
}

# Run the Shiny app
shinyApp(ui, server)
}
\seealso{
\code{\link{mod_file_picker_server}} for the corresponding server
 part of the module.
}
