% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_save_file_generic.R
\name{mod_save_file_generic_server}
\alias{mod_save_file_generic_server}
\title{Generic module server function for saving files for export to Platform}
\usage{
mod_save_file_generic_server(id, reac_vals, sbg_directory_path = "/sbgenomics")
}
\arguments{
\item{id}{Module's ID.}

\item{reac_vals}{Reactive values list containing mandatory fields:
\itemize{
 \item `FUN`- Function for creating a file for saving, i.e `write.table`,
    `save`, `write`, `write_json`, `write_xml`, `SaveH5Seurat` etc.
 \item `args`- List of function arguments for the provided FUN.
 \item `filename`- File name.
 \item `extension`- Expected file extension. Please provide expected file
    extension if exists, in order to properly validate the existence of the
    file with the same name and extension.
 \item `overwrite`- Boolean. Overwrite existing file with the same name.
}}

\item{sbg_directory_path}{Path to the mounted `sbgenomics` directory
containing `project-files`, `output-files` and `workspace` sub-directories
on the instance.
These directories are expected to exist on the instance where the app would
run. For the purposes of testing your app locally, you can create a mock
directory `sbgenomics` with the same structure - containing sub-directories
`project-files`, `output-files` and `workspace` and populate with test files
mimicking the project's file structure on the Platform.}
}
\value{
No value. Use in UI & server of shiny application.
}
\description{
A server side of a shiny Module that allows users to save files
 generated within the Shiny app and export them to the project on the
 Seven Bridges Platform.
}
\examples{
library(shiny)
library(magrittr)
library(sbShinyModules)
library(reactable)
library(jsonlite)

# App's  UI
ui <- fluidPage(
  titlePanel("Save data for export to Platform - Module Demo"),
  sidebarLayout(
    sidebarPanel(
      fluidRow(
        h3("Export txt/csv to Platform"),
        br(),
        textInput("file_name", label = "Set file name", width = "100\%"),
        fluidRow(
          column(
            width = 6,
            selectInput("separator",
              label = "Set separator",
              choices = c("comma", "tab", "new line", "semicolon")
            )
          ),
          column(
            width = 6,
            selectInput("extension",
              label = "Set extension",
              choices = c("txt", "csv")
            )
          )
        ),
        fluidRow(
          column(
            width = 6,
            radioButtons("overwrite",
              label = "Overwrite existing file?",
              choiceNames = c("yes", "no"),
              choiceValues = c("TRUE", "FALSE")
            )
          )
        ),
        fluidRow(
          column(
            width = 4,
            sbShinyModules::mod_save_file_generic_ui(
              id = "file_exporter",
              save_button_title = "Save txt/csv"
            )
          )
        )
      ),
      fluidRow(
        h3("Export json to Platform"),
        br(),
        textInput("file_name_json", label = "Set file name", width = "100\%"),
        fluidRow(
          column(
            width = 4,
            sbShinyModules::mod_save_file_generic_ui(
              id = "json_file_exporter",
              save_button_title = "Save json"
            )
          )
        )
      ),
      fluidRow(
        h3("Export RDS object to Platform"),
        br(),
        textInput("file_name_rds", label = "Set file name", width = "100\%"),
        fluidRow(
          column(
            width = 4,
            sbShinyModules::mod_save_file_generic_ui(
              id = "rds_file_exporter",
              save_button_title = "Save RDS"
            )
          )
        )
      )
    ),
    mainPanel(
      reactable::reactableOutput("table_data")
    )
  )
)

# App's Server Logic
server <- function(input, output, session) {
  output$table_data <- reactable::renderReactable({
    reactable::reactable(iris,
      onClick = "select",
      filterable = TRUE,
      searchable = TRUE,
      resizable = TRUE,
      defaultPageSize = 10
    )
  })

  helper_rv_table <- reactiveValues(
    FUN = write.table,
    args = list(x = iris, quote = FALSE, row.names = FALSE, col.names = TRUE),
    filename = NULL,
    extension = NULL,
    overwrite = FALSE
  )

  observe(
    helper_rv_table$filename <- as.character(input$file_name)
  )
  observe(
    helper_rv_table$args[["sep"]] <- switch(input$separator,
      "comma" = ",",
      "tab" = "\t",
      "new line" = "\n",
      "semicolon" = ";"
    )
  )
  observe(
    helper_rv_table$extension <- as.character(input$extension)
  )
  observe(
    helper_rv_table$overwrite <- as.logical(input$overwrite)
  )

  # Call the file exporter module
  sbShinyModules::mod_save_file_generic_server(
    id = "file_exporter",
    reac_vals = helper_rv_table,
    sbg_directory_path = system.file(
      "tests", "testthat", "sbgenomics_test",
      package = "sbShinyModules"
    )
  )

  helper_rv_json <- reactiveValues(
    FUN = write,
    args = list(
      x = jsonlite::toJSON(x = iris, dataframe = "rows", pretty = TRUE)
    ),
    filename = NULL,
    extension = "json",
    overwrite = TRUE
  )

  observe(
    helper_rv_json$filename <- as.character(input$file_name_json)
  )

  # Call the file exporter module
  sbShinyModules::mod_save_file_generic_server(
    id = "json_file_exporter",
    reac_vals = helper_rv_json,
    sbg_directory_path = system.file(
      "tests", "testthat", "sbgenomics_test",
      package = "sbShinyModules"
    )
  )

  helper_rv_rds <- reactiveValues(
    FUN = saveRDS,
    args = list(
      object = iris
    ),
    filename = NULL,
    extension = "RDS",
    overwrite = TRUE
  )

  observe(
    helper_rv_rds$filename <- as.character(input$file_name_rds)
  )

  # Call the file exporter module
  sbShinyModules::mod_save_file_generic_server(
    id = "rds_file_exporter",
    reac_vals = helper_rv_rds,
    sbg_directory_path = system.file(
      "tests", "testthat", "sbgenomics_test",
      package = "sbShinyModules"
    )
  )
}

# Run the Shiny app
shinyApp(ui, server)
}
