% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_save_plot_to_export.R
\name{mod_save_plot_to_export_server}
\alias{mod_save_plot_to_export_server}
\title{Save `plot` for export module server function}
\usage{
mod_save_plot_to_export_server(
  id,
  plot_reactVals,
  output_formats = get_golem_config("PLOT_EXPORT_SUPPORTED_EXT"),
  module_title = "Save plot for export",
  sbg_directory_path = "/sbgenomics",
  btns_div_width = 12
)
}
\arguments{
\item{id}{Module's ID.}

\item{plot_reactVals}{A `reactiveValues` variable with a slot `plot`
 containing an object created with `recordPlot()` function you would like to
 save. Please, check out our example app in inst/demos folder -
 `plot_exporter_demo_app.R` for demonstrating the usage of this module. \cr
 Example:
 ```{r}
 helper_reactive <- reactiveValues(
  plot = NULL
 )

 # Some observer or eventReactive where you create your plot
 # and then recordPlot() at the end:
 plot_output <- eventReactive(input$bins, {
 bins <- input$bins + 1

 # Draw the histogram with the specified number of bins
 faithful[, 2] %>%
   hist(
     breaks = seq(min(.),
                  max(.),
                  length.out = bins
     ),
     col = "darkgray",
     border = "white",
     main = "Geyser eruption duration"
   )
 helper_reactive$plot <- recordPlot()
 })

 output$distPlot <- renderPlot({
   plot_output()
 })
 ```}

\item{output_formats}{Output formats supported. Can be a subset of:
"png", "pdf", "svg", "jpeg", "bmp", or "tiff".}

\item{module_title}{Title (top left corner) of a modal (popup window) with
settings.}

\item{sbg_directory_path}{Path to the mounted `sbgenomics` directory
containing `project-files`, `output-files` and `workspace` sub-directories
on the instance.
These directories are expected to exist on the instance where the app would
run. For the purposes of testing your app locally, you can create a mock
directory `sbgenomics` with the same structure - containing sub-directories
`project-files`, `output-files` and `workspace` and populate with test files
mimicking the project's file structure on the Platform.}

\item{btns_div_width}{Width of the `div()` containing block with buttons for
saving plots in the pop-up modal dialogue. Suitable to update when having
one or two buttons out of supported set. Default is 12.}
}
\value{
No value. Use in UI & server of shiny application.
}
\description{
A server side of a shiny Module that allows users to save plots
 generated within the Shiny app and export them to the project on the
 Seven Bridges Platform.
}
\examples{
library(shiny)
library(magrittr)
library(sbShinyModules)

# App's  UI
ui <- fluidPage(
  titlePanel("Save plots for export to Platform - Module Demo"),
  sidebarLayout(
    sidebarPanel(
      sliderInput("bins",
        "Number of bins:",
        min = 1,
        max = 50,
        value = 30
      )
    ),
    mainPanel(
      plotOutput("distPlot"),
      fluidRow(
        h3("Export Plot to Platform"),
        br(),
        sbShinyModules::mod_save_plot_to_export_ui(
          id = "plot_exporter"
        )
      )
    )
  )
)

# App's Server Logic
server <- function(input, output, session) {
  # Create reactive values list with plot field
  helper_reactive <- reactiveValues(
    plot = NULL
  )

  plot_output <- eventReactive(input$bins, {
    bins <- input$bins + 1

    # Draw the histogram with the specified number of bins
    faithful[, 2] \%>\%
      hist(
        breaks = seq(min(.),
          max(.),
          length.out = bins
        ),
        col = "darkgray",
        border = "white",
        main = "Geyser eruption duration"
      )
    helper_reactive$plot <- recordPlot()
  })

  output$distPlot <- renderPlot({
    plot_output()
  })

  # Call the plot exporter module
  sbShinyModules::mod_save_plot_to_export_server(
    id = "plot_exporter",
    plot_reactVals = helper_reactive,
    module_title = "Save plot to Platform",
    sbg_directory_path = system.file(
      "tests", "testthat", "sbgenomics_test",
      package = "sbShinyModules"
    ),
    btns_div_width = 12
  )
}

# Run the Shiny app
shinyApp(ui, server)
}
